enlarging chunksize to 512KB but does not help improve perf on amd; fixing size overflow bugs

+++ rdma/util_rdma.h
@@ -87,7 +87,7 @@ static inline int util_rdma_get_link_speed_from_ibv_speed(int active_speed,
  * @brief Buffer pool for work request extension.
  */
 class WrExBuffPool : public BuffPool {
-  static constexpr uint32_t kWrSize = sizeof(struct wr_ex);
+  static constexpr size_t kWrSize = sizeof(struct wr_ex);
   static constexpr uint32_t kNumWr = 4096;
   static_assert((kNumWr & (kNumWr - 1)) == 0, "kNumWr must be power of 2");
 
@@ -113,7 +113,7 @@ struct CQEDesc {
 
 class CQEDescPool : public BuffPool {
  public:
-  static constexpr uint32_t kDescSize = sizeof(struct CQEDesc);
+  static constexpr size_t kDescSize = sizeof(struct CQEDesc);
   static constexpr uint32_t kNumDesc = 4 * 65536;
   static_assert((kNumDesc & (kNumDesc - 1)) == 0,
                 "kNumDesc must be power of 2");
@@ -199,8 +199,7 @@ struct __attribute__((packed)) retr_chunk_hdr {
  */
 class RetrChunkBuffPool : public BuffPool {
  public:
-  static constexpr uint32_t kRetrChunkSize =
-      kChunkSize + sizeof(retr_chunk_hdr);
+  static constexpr size_t kRetrChunkSize = kChunkSize + sizeof(retr_chunk_hdr);
   static constexpr uint32_t kNumChunk = 4096;
   static_assert((kNumChunk & (kNumChunk - 1)) == 0,
                 "kNumChunk must be power of 2");
@@ -216,7 +215,7 @@ class RetrChunkBuffPool : public BuffPool {
  */
 class RetrHdrBuffPool : public BuffPool {
  public:
-  static constexpr uint32_t kHdrSize = sizeof(struct retr_chunk_hdr);
+  static constexpr size_t kHdrSize = sizeof(struct retr_chunk_hdr);
   static constexpr uint32_t kNumHdr = 1024;
   static_assert((kNumHdr & (kNumHdr - 1)) == 0, "kNumHdr must be power of 2");
 
@@ -230,8 +229,8 @@ class RetrHdrBuffPool : public BuffPool {
  */
 class CtrlChunkBuffPool : public BuffPool {
  public:
-  static constexpr uint32_t kPktSize = 36;
-  static constexpr uint32_t kChunkSize = kPktSize * kMaxBatchCQ;
+  static constexpr size_t kPktSize = 36;
+  static constexpr size_t kChunkSize = kPktSize * kMaxBatchCQ;
   static constexpr uint32_t kNumChunk = 65536;
   static_assert((kNumChunk & (kNumChunk - 1)) == 0,
                 "kNumChunk must be power of 2");
@@ -1128,9 +1127,9 @@ struct pair_hash {
 // Shared IO context for each UCCL engine.
 class SharedIOContext {
  public:
-  constexpr static int kRetrMRSize =
+  constexpr static size_t kRetrMRSize =
       RetrChunkBuffPool::kRetrChunkSize * RetrChunkBuffPool::kNumChunk;
-  constexpr static int kCtrlMRSize =
+  constexpr static size_t kCtrlMRSize =
       CtrlChunkBuffPool::kChunkSize * CtrlChunkBuffPool::kNumChunk;
   SharedIOContext(int dev) {
     auto context = RDMAFactory::get_factory_dev(dev)->context;
